zam zam

+++ lodger/index.js
@@ -3,6 +3,8 @@ import { definitii, helperCautare } from './definitii'
 import colectii  from './db/collections'
 
 let initializat
+let _db
+const api = {}
 const { NODE_ENV } = process.env
 
 /**
@@ -16,21 +18,43 @@ export default class Lodger {
     if (!context || typeof context._db === 'undefined') {
       throw new Error('DB neinitializat')
     }
+    _db = context._db
 
-    this._db = context._db
-
-    /**
-     * props accesibile doar pentru teste
-     */
-    if (NODE_ENV === 'test') {
-      this.definitii = definitii
-      this.helperCautare = helperCautare
-      // this.notifica = notifica
+    // api-ul
+    for (const [singular, plural] of definitii) {
+      const colectie = _db[plural]
+  
+      Object.defineProperties(this, {
+        [singular]: {
+          value: { 
+            ... colectie._methods,
+            adauga: colectie.insert.bind(colectie)
+          }
+        },
+        [plural]: {
+          get () {
+            if (!arguments) {
+              return 'default search - limit etc'
+            }
+            // getterul custom cu criteri
+            // eg. lodger.asociatii({ querycautare })
+            return async criteriu => {
+              return await colectie.find(criteriu).exec()
+            }
+          }
+        }
+      })
     }
 
     /* Indicator initializare - flag */
     initializat = true
+    console.dir('api', api)
+
     return this
+  }
+
+  get _db () {
+    return _db
   }
 
   get initializat () {
@@ -50,7 +74,7 @@ export default class Lodger {
   }
 
   async destroy () {
-    if (!this.initializat) return
+    if (!this._db || !this.initializat) return
     await this._db.destroy()
   }
 }
